requirements.txt
openpyxl
Pillow
